+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.8] - 2026-01-30
+
+### Added
+
+- Folder management in Settings > Folders
+  - Add new folders with "Add Folder" button
+  - Rename folders inline with edit button
+  - Delete folders with option to move feeds to another folder or delete them
+  - Shows feed count per folder
+
 ## [0.4.7] - 2026-01-30
 
 ### Added